requirements.txt
Django==2.0.7
django-appconf==1.0.2
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-webpack-loader==0.6.0
Pillow==5.2.0
pytz==2018.5
